packages/nlp-service/pyproject.toml
[project]
name = "nlp-service"
version = "1.0.0"
description = "NLP Service for PDF Quiz Generator - Text extraction, chunking, and question generation"
authors = [{ name = "PDF Quiz Gen Team" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "pymupdf==1.23.8",
    "spacy==3.7.2",
    "python-multipart==0.0.6",
    "httpx==0.26.0",
    "redis==5.0.1",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "httpx>=0.26.0",
    "ruff==0.1.9",
    "mypy==1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
